fixed issues

+++ src/models/Bracket.js
@@ -29,12 +29,12 @@ const Bracket = sequelize.define(
       defaultValue: 0,
     },
     minRating: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(4, 2),
       allowNull: false,
       defaultValue: 0,
     },
     maxRating: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(4, 2),
       allowNull: false,
       defaultValue: 0,
     },
